moved config.logging_helper to logger and tests
