+++ astropy/config/__init__.py
@@ -10,5 +10,3 @@ from .paths import *
 from .configuration import *
 from .data import *
 from .affiliated import *
-from .logging_helper import *
-del log  # we remove the log here because it is imported in astropy/__init__.py
